ci: refactor to use pyproject.toml for dependencies, collapse scenario and unit tests
