+++ tests/unit/test_ingress_per_app.py
@@ -17,8 +17,8 @@ from charms.traefik_k8s.v2.ingress import (
 from ops import CharmBase, Framework
 from scenario import Context, Model, Mount, Relation, State
 
-from tests.scenario._utils import create_ingress_relation
-from tests.scenario.conftest import MOCK_LB_ADDRESS
+from conftest import MOCK_LB_ADDRESS
+from tests.unit._utils import create_ingress_relation
 
 
 @pytest.mark.parametrize(
