+++ tests/unit/test_ingress_tls.py
@@ -6,7 +6,7 @@ import pytest
 import yaml
 from scenario import Container, ExecOutput, Mount, Relation, State
 
-from tests.scenario._utils import _render_config, create_ingress_relation
+from tests.unit._utils import _render_config, create_ingress_relation
 
 
 def _create_tls_relation(*, app_name: str, strip_prefix: bool, redirect_https: bool):
