+++ tests/unit/test_ingress_v1_backwards_compat/test_ingress_tls_v1.py
@@ -5,7 +5,7 @@ import pytest
 import yaml
 from scenario import Container, Mount, Relation, State
 
-from tests.scenario._utils import _render_config
+from tests.unit._utils import _render_config
 
 
 def _create_ingress_relation(
