+++ tests/unit/test_tracing_integration.py
