+++ tox.ini
@@ -4,15 +4,17 @@
 [tox]
 skipsdist=True
 skip_missing_interpreters = True
-envlist = lint, static-{charm, lib}, unit, scenario, interface
+envlist = lint, static, unit
 
 [vars]
 src_path = {toxinidir}/src
-tst_path = {toxinidir}/tests
+tests_path = {toxinidir}/tests
 lib_path = {toxinidir}/lib/charms/traefik_k8s
 all_path = {[vars]src_path} {[vars]tst_path} {[vars]lib_path}
+uv_flags = --frozen --isolated  --all-groups
 
 [testenv]
+allowlist_externals = uv
 setenv =
   JUJU_VERSION=3.0.3 # fix ops borkiness
   PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
@@ -20,86 +22,54 @@ setenv =
   PY_COLORS=1
 passenv =
   PYTHONPATH
+  CHARM_PATH
   CHARM_BUILD_DIR
   MODEL_SETTINGS
+usedevelop = false
+basepython = python3
+
+[testenv:lock]
+description = Update uv.lock with the latest deps
+commands =
+    uv lock --upgrade --no-cache
 
 [testenv:fmt]
 description = Apply coding style standards to code
-deps =
-    black
-    ruff
 commands =
-    ruff check --fix {[vars]all_path}
-    black {[vars]all_path}
+    uv run {[vars]uv_flags} ruff check --fix-only {[vars]all_path}
 
 [testenv:lint]
 description = Check code against coding style standards
-deps =
-    black
-    ruff
-    codespell
 commands =
-    codespell {[vars]lib_path}
-    codespell .
-    ruff check {[vars]all_path}
-    black --check --diff {[vars]all_path}
+    uv run {[vars]uv_flags} ruff check {[vars]all_path}
+    uv run codespell {[vars]lib_path}
+    uv run codespell .
 
 [testenv:unit]
 description = Run unit tests
-deps =
-    pytest
-    pytest-subtests
-    coverage[toml]
-    ipdb
-    -r{toxinidir}/requirements.txt
+allowlist_externals=
+    {[testenv]allowlist_externals}
+    /usr/bin/env
 commands =
-    coverage run --source={[vars]src_path} \
-        -m pytest -vv --tb native --log-cli-level=INFO -s {[vars]tst_path}/unit {posargs}
-    coverage report
+    uv run {[vars]uv_flags} coverage run --source={[vars]src_path} -m pytest \
+        {[vars]tests_path}/unit {posargs}
+    uv run {[vars]uv_flags} coverage report
 
 [testenv:integration]
 description = Run integration tests
-deps =
-    pytest==8.2.2
-    pytest-asyncio==0.21.2
-    pytest-operator
-    juju
-    # fix for https://github.com/jd/tenacity/issues/471
-    tenacity==8.3.0
-    sh
-    minio
-    -r{toxinidir}/requirements.txt
 commands =
-    pytest -v --tb native --log-cli-level=INFO -s {[vars]tst_path}/integration {posargs}
-;    pytest -v --tb native --log-cli-level=INFO -s {posargs}
+    uv run {[vars]uv_flags} pytest --exitfirst {[vars]tests_path}/integration {posargs}
 
-[testenv:static-{charm,lib}]
-description = Static code checking
-deps =
-    pyright==1.1.318
-    -r{toxinidir}/requirements.txt
+[testenv:static]
+description = Run static checks
+allowlist_externals =
+  {[testenv]allowlist_externals}
+  /usr/bin/env
 commands =
-    charm: pyright --pythonversion 3.8 {[vars]src_path}
-    lib: pyright --pythonversion 3.8 {[vars]lib_path}
-    lib: /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
-allowlist_externals = /usr/bin/env
-
-[testenv:scenario]
-description = Scenario tests
-deps =
-    pytest
-    ops-scenario<7.0.0
-    -r{toxinidir}/requirements.txt
-commands =
-    pytest -v --tb native {[vars]tst_path}/scenario --log-cli-level=INFO -s {posargs}
+    uv run {[vars]uv_flags} pyright {[vars]all_path}
+    /usr/bin/env sh -c 'for m in $(git diff main --name-only {[vars]lib_path}); do if ! git diff main $m | grep -q "+LIBPATCH\|+LIBAPI"; then echo "You forgot to bump the version on $m!"; exit 1; fi; done'
 
 [testenv:interface]
 description = Run interface tests
-deps =
-    pytest
-    ops-scenario~=6.0
-    pytest-interface-tester > 0.3
-    -r{toxinidir}/requirements.txt
 commands =
-    # todo uncomment once scenario v7 migration on interface tester is complete
-    # pytest -v --tb native {[vars]tst_path}/interface --log-cli-level=INFO -s {posargs}
+    uv run pytest -v --tb native {[vars]tst_path}/interface --log-cli-level=INFO -s {posargs}
